Adding files for the application.
Currently it is setup as a sunrise sunset calculator, accesing third party resources, utilizing promise chains,
capitilazing off of Angular with one service implemented so far which calles the third party apis.  Then there
is the deployd mongodb backend which the controller currently speaks to.  This is a rough frame, dialed in already
for code standards, and cleaniliness!  Its pretty much very useful scaffolding for a turn key deployd application
that uses angular for its front end and middle tier.

The styles implemented so far are pretty much bootstrap, with an aim for minimal jquery usage.  The angular app's
jqlite has already been replaced to using a full version of jquery.

Some integration tests are included in the tests directory, and are for Selenium IDE.